requirements.txt
torch==2.7.0
torchvision==0.22.0
datasets==2.21.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
pandas>=2.2.0
PyYAML>=6.0.0
tqdm>=4.66.0
Pillow>=10.0.0
